cleanup cache step
