+++ src/main/java/io/jenkins/plugins/pipeline/cache/CacheStep.java
@@ -1,10 +1,13 @@
 package io.jenkins.plugins.pipeline.cache;
 
+import java.io.IOException;
+import java.io.PrintStream;
 import java.io.Serializable;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.Set;
 
+import org.jenkinsci.plugins.workflow.steps.BodyExecutionCallback;
+import org.jenkinsci.plugins.workflow.steps.GeneralNonBlockingStepExecution;
 import org.jenkinsci.plugins.workflow.steps.Step;
 import org.jenkinsci.plugins.workflow.steps.StepContext;
 import org.jenkinsci.plugins.workflow.steps.StepDescriptor;
@@ -12,26 +15,49 @@ import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 
 import hudson.Extension;
+import hudson.FilePath;
 import hudson.model.TaskListener;
+import io.jenkins.plugins.pipeline.cache.agent.BackupCallable;
+import io.jenkins.plugins.pipeline.cache.agent.RestoreCallable;
 
 /**
- * Provides a pipeline step to cache files located on the build agent. Before the step gets executed, the files are restored from S3
- * and afterwards the state is sored in S3. See README.md for more details.
+ * Handles 'cache' step executions.<br><br>
+ * How it works?
+ * <ol>
+ *     <li>cache gets restored by the given key to the given path (if key exists or one of the restoreKeys matches)</li>
+ *     <li>inner-step gets executed</li>
+ *     <li>backup of the path gets created (only if inner-step was successful and if the key not already exists)</li>
+ * </ol>
+ * Note: When a cache gets restored then a list of keys (key is the first one followed by the restoreKeys) is used to find a matching key
+ * . See {@link io.jenkins.plugins.pipeline.cache.s3.CacheItemRepository#findRestoreKey(String, String...)} for more details.
  */
 public class CacheStep extends Step implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
+    /**
+     * (required) Path to the folder (absolute or relative to the workspace) which should be cached (e.g. <i>$HOME/.m2/repository</i>).
+     */
     private final String path;
-    private final String key;
-
-    private String[] restoreKeys;
-    private String filter;
 
     /**
-     * @param path Absolute path to the folder you want to cache (e.g. $HOME/.m2/repository)
-     * @param key Identifier assigned to this cache (e.g. maven-${hashFiles('**&#47;pom.xml')})
+     * (required) The unique identifier which is assigned to this cache (e.g. <i>maven-${hashFiles('**&#47;pom.xml')}</i>).
      */
+    private final String key;
+
+    /**
+     * (optional) Additional keys which are used when the cache gets restored and the key doesn't exist (e.g. <i>maven-,
+     * maven-4f98f59e877ecb84ff75ef0fab45bac5</i>).
+     */
+    @DataBoundSetter
+    private String[] restoreKeys;
+
+    /**
+     * (optional) Glob pattern to filter the path (e.g. <i>**&#47;*.java</i> includes java files only).
+     */
+    @DataBoundSetter
+    private String filter;
+
     @DataBoundConstructor
     public CacheStep(String path, String key) {
         this.path = path;
@@ -41,38 +67,6 @@ public class CacheStep extends Step implements Serializable {
     @Override
     public CacheStepExecution start(StepContext context) throws Exception {
         return new CacheStepExecution(context, this);
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public String getKey() {
-        return key;
-    }
-
-    public String[] getRestoreKeys() {
-        return restoreKeys == null ? null : Arrays.copyOf(restoreKeys, restoreKeys.length);
-    }
-
-    /**
-     * @param restoreKeys Additional keys used to restore the Cache (e.g. maven-, maven-4f98f59e877ecb84ff75ef0fab45bac5)
-     */
-    @DataBoundSetter
-    public void setRestoreKeys(String[] restoreKeys) {
-        this.restoreKeys = restoreKeys == null ? null : Arrays.copyOf(restoreKeys, restoreKeys.length);
-    }
-
-    public String getFilter() {
-        return filter;
-    }
-
-    /**
-     * @param filter Ant file pattern mask, like <b>**&#47;*.java</b> which is applied to the path.
-     */
-    @DataBoundSetter
-    public void setFilter(String filter) {
-        this.filter = filter;
     }
 
     @Extension
@@ -92,4 +86,53 @@ public class CacheStep extends Step implements Serializable {
             return true;
         }
     }
+
+    private static class CacheStepExecution extends GeneralNonBlockingStepExecution {
+
+        private static final long serialVersionUID = 1L;
+
+        private final transient PrintStream logger;
+        private final CacheStep step;
+        private final CacheConfiguration config;
+
+        public CacheStepExecution(StepContext context, CacheStep step) throws IOException, InterruptedException {
+            super(context);
+            this.step = step;
+            this.logger = context.get(TaskListener.class).getLogger();
+            this.config = CacheConfiguration.get();
+        }
+
+        @Override
+        public boolean start() throws Exception {
+            FilePath workspace = getContext().get(FilePath.class);
+            FilePath path = workspace.child(step.path);
+
+            // restore existing cache
+            path.act(new RestoreCallable(config, step.key, step.restoreKeys)).printInfos(logger);
+
+            // execute inner-step and save cache afterwards
+            getContext().newBodyInvoker().withCallback(new BodyExecutionCallback() {
+                @Override
+                public void onSuccess(StepContext context, Object result) {
+                    try {
+                        path.act(new BackupCallable(config, step.key, step.filter)).printInfos(logger);
+                    } catch (Exception x) {
+                        context.onFailure(x);
+                        return;
+                    }
+                    context.onSuccess(result);
+                }
+
+                @Override
+                public void onFailure(StepContext context, Throwable t) {
+                    logger.println("Cache not saved (inner-step execution failed)");
+                    context.onFailure(t);
+                }
+            }).start();
+
+            return false;
+        }
+
+    }
+
 }
